Move the API's search code into its own module
